requirements.txt
keyboard
pyperclip
anthropic
python-dotenv
customtkinter
pillow
pystray
screeninfo